[BM算法]: 初始化获取坏字符xi位置数组

+++ leet-code/src/main/java/com/learning/algoithm/search/BMSearch.java
@@ -19,7 +19,7 @@ public class BMSearch implements StringSearch {
             bc[i] = -1;
         }
         for (int i = 0; i < searchStrChars.length; i++) {
-            int ascii = (int) searchStrChars[i];
+            int ascii = searchStrChars[i];
             bc[ascii] = i;
         }
